// src/app/pages/items/models/list-all-subjects.model.ts
export class ListAllSubjects {
  
  name: string;
  subjectId: string;
  totalPassages: number;
  subjectCode: string;
  totalItems: number;
  awaitingModeration: number;
  totalDrafts: number;
  totalUsed: number;
  inRecycle: number;
  totalPassagesDraft: number;
  totalPassagesPublished: number;
  moderationId: string;
  totalPublished: number;
  totalPassagesAwaitingModeration: number;
}